CHORE: Redis dev 환경 세팅 후 실행 성공 - 환경 세팅 진짜 마지막

See merge request s10-blockchain-contract-sub2/S10P22D210!79

+++ backend/src/main/java/org/ssafy/d210/walk/dto/response/ThisWeekExerciseResponseDto.java
@@ -9,5 +9,5 @@ import java.time.LocalDate;
 @Builder
 public class ThisWeekExerciseResponseDto {
     private LocalDate timeStamp;
-    private Integer steps;
+    private Long steps;
 }
